adding unit test, cleaning/reorganizing, fixing the non-cxx11 version

+++ include/common/defs.h
@@ -138,10 +138,10 @@ using  static_short=boost::mpl::integral_c<short_t,N>;
 template<ushort_t N>
 using  static_ushort=boost::mpl::integral_c<ushort_t,N>;
 #else
-typedef long int                     int_t;
+typedef int                     int_t;
 typedef int                     short_t;
-typedef unsigned long int            uint_t;
-typedef unsigned int            ushort_t;
+typedef int                     uint_t;
+typedef int                     ushort_t;
 template<int_t N>
 struct static_int : boost::mpl::integral_c<int_t,N>{
     typedef boost::mpl::integral_c<int_t,N> type;
